testapi: Allow distri class provide default for die_on_timeout
script_run warns if die_on_timeout is not provided (regardless if is 0
or 1). Adjusting all the usages to always use the argument is
unnecessary boilerplate, so let the distri class provide the default.
This way, test developers can set it everywhere without actually
changing every single call. It is still possible to override it
per-call.