Force full screen update shortly before check screen would fail
Sometimes we are left with a distorted VNC screen despite no stall is
detected. This problem is sometimes even reproducible when running the
fullstack test (fail rate is 5 % on my system), see
https://progress.opensuse.org/issues/104971.

This change forces a full (non-incremental) update of the VNC screen
shortly before a check screen would fail to workaround the issue.