Avoid non-incremental screen updates in `localXvnc` console
It apparently breaks in some cases, see
https://progress.opensuse.org/issues/106017 where the VNC connection is
attempted to be re-established as it is considered stale (but it fails).
The additional non-incremental update requests have maybe an impact on the
stale VNC detection making it more aggressive. Note that when watching the
test referenced in the ticket with the non-incremental screen update
requests commented out it passes but the screen is not updated for a while.

This is basically a revert of 944b3626c89b47baf7162013a85f2d54b7858131 for
localXvnc.