CI: Add automatic review checklist based on changed file paths
One important factor for myself and apparently also other reviewers is
that we keep forgetting or overlooking certain things in reviews because
we review in multiple different projects and for each we need to look
for different things, e.g. in openQA I can usually trust the automated
test assuming that there are according tests added and code coverage is
fine, for os-autoinst I would need to carefully try myself or ask for
tests, for openQA test distributions I would need to look for according
openQA verification jobs.

With this commit pull requests introducing changes to any backend file
as well as testapi.pm will receive an according review checklist so that
reviewers as well as reviewees do not forget such changes need to be
tested accordingly or API version changes need to be applied.

Related progress issue: https://progress.opensuse.org/issues/107005