Remove obsolete python2 support in crop.py
crop.py was the only python-script that was still using a python2
shebang although that was anyway already replaced with a python3 shebang
in the openSUSE spec file. As we never pulled in necessary dependencies
in our spec file we can also simply change that to remove obsolete
python2 support. Then we also don't need the override rules in the spec
file anymore.