Add error handling for uninitialized isotovideo socket
When trying to call test API functions from schedule code, e.g. the
main.pm one would get an error message like:

```
myjsonprc: called on undefined file descriptor at
/usr/lib/os-autoinst/myjsonrpc.pm line 40.
```

which is not very helpful to beginners. The problem is that test API
functions which only work as expected from within test modules are
called by users in unexpected cases, e.g. from scheduling code like a
main.pm file. In this case the isotovideo socket was not yet initialized
and we would try to execute a command on the worker, not the system
under test, for example a virtual machine.

This commit introduces an explicit message with a hint to the user what
to check in case they try to call tests in the wrong way where we would
end up calling the send_json function within an uninitialized object.

The problem was originally discovered when supporting a user receiving
such unexpected errors.