t: Ensure to show exception details when python test modules fail
In case our inline python code would fail a python exception would be
thrown which would abort the perl test but as we have caught the stderr
output at this time already the actual error details would be lost and
vanished. This commit introduces an explicit "lives_ok" check which
catches if there is an exception and provides the complete output in all
cases of error.