myjsonrpc: Don't limit how often select may return EINTR
select returns EINTR if the wait was interrupted by any signal. The previous
code errored out if this happened more than five times, which is not wrong
and not unlikely to happen. The limit doesn't really have a purpose anyway,
so just remove it.

Related ticket: https://progress.opensuse.org/issues/107710