Fix qemu test to actually wait for the process to be stopped
Just calling ->stop does not guarantee that the forked process stops in
time for the cleanup code to create the state file for the test code. A
switch to ->wait_stop should make the test more reliable. If that does
not resolve the issue completely, we might have to hook into an event to
manually wait for the process to stop.

Progress: https://progress.opensuse.org/issues/108443