Hide secrets in all log_call invocations
My first idea was to encrypt the text and then decrypt it when it is needed for
the call.

However the concern was about the logs and i found that the only thing that it
needs to be done
is to update the log_call, because this is what is called from all the backends.
in this way nothing else seems to need to change (in sense of manipulate the
_text_ before pass to log_call) in testapi or anywhere else.

So i have removed condition in testapi but the question is do you need the `log_call` there.
Seems like an extra line in os-autoinst

```
[2022-03-28T12:34:13.170200+02:00] [debug] tests/console/aaa_base.pm:31 called opensusebasetest::select_serial_terminal -> lib/opensusebasetest.pm:1288 called testapi::select_console -> lib/susedistribution.pm:754 called serial_terminal::login -> lib/serial_terminal.pm:142 called testapi::type_password
[2022-03-28T12:34:13.170414+02:00] [debug] <<< testapi::type_string(text="*****", secret=1, max_interval=100)
[2022-03-28T12:34:13.171423+02:00] [debug] <<< consoles::serial_screen::type_string(max_interval=100, cmd="backend_type_string", text="*****", secret=1, json_cmd_token="nJFsVgxh")
```

Signed-off-by: ybonatakis <ybonatakis@suse.com>