Avoid JSON::PP::Boolean warnings
When Cpanel::JSON::XS is loaded before JSON::PP, it will
result in warnings

    Subroutine JSON::PP::Boolean::(0+ redefined at .../overload.pm
    ...

This is because Cpanel::JSON::XS creates the same overloads
as JSON::PP. This is handled in newer versions of JSON::PP, so
it will go away at some point, but until then it is always confusing to see
the warnings.
Loading it in isotovideo at the beginning first will avoid the warning.

Issue: https://progress.opensuse.org/issues/90371