Add githook-perltidy as pre-commit git hook
This PR adds `App::githook::perltidy` in the os-autoinst dependencies.
With this in place contributors will ensure that the coding style is
sustained and remains clean and tidy.

To install the script the only requirement is the existence of `.perltidyrc`
which this repo already has. For other cases take a look on
https://metacpan.org/dist/App-githook-perltidy/view/bin/githook-perltidy#githook-perltidy-pre-commit

Signed-off-by: ybonatakis <ybonatakis@suse.com>