Apply VNC timeout also for reading/writing
This change applies the VNC timeout introduced in the previous commit
also to receiving/sending.

In my tests the timeout specified so far only via the constructor is not
effective. I suppose it is only used for connect/accept¹. So this commit
manually sets the socket option to apply the timeout for receiving/sending
as well.

I tested it on a production worker and it works. Even despite not handling
`EAGAIN` explicitly the VNC processing is even able to continue where it
was previously hanging letting the test pass.

---

¹ See https://stackoverflow.com/questions/8284243/how-do-i-set-so-rcvtimeo-on-a-socket-in-perl