Fix perl 5.36 warnings on use of @_ in functions with signature
When running the test suite with perl 5.36 we get several warnings
about use (explicit or implict) of `@_` in functions with
signatures being 'experimental'.

These fixes were all suggested by tinita (thanks). In some cases
we can just use the function arguments, in three cases that fill
buffers we have to drop the signatures, as you cannot assign to
`@_` without triggering the warning when using signatures, and we
don't see an obvious way to have these functions do anything else.
In console_proxy we also drop the signature; we considered calling
`$AUTOLOAD` with the args instead, but @kraih thinks that will
cause problems with stack traces.

Signed-off-by: Adam Williamson <awilliam@redhat.com>