Fix error from stricter git permission checks
git version 2.35 introduced strict checks on directory ownership
refusing to operate on directories not owned by the user calling git.
In the case of openQA tests not running from openQA worker cache
synchronized test distributions this caused the git revision of test
distribution directories to be not parsed and ending up as 'UNKNOWN',
our fallback string.

This commit fixes this problem by adding the requested git directory to
the list of "safe directories" in a temporary user's global git config.
This of course means that the user account must be able to write to the
according config file so this is why we use a temporary home directory.

Related progress issue: https://progress.opensuse.org/issues/113030