Add back support for reading layered test distribution's git versions
In c9631656 "Allow running tests (invoking isotovideo) without Git" a
check for the git file/directory was introduced. This broke reading out
the git hash in case of test distributions or needle directories that
are not in the root directory of git repositories.

This commit fixes this by not trying to find the git directory but
instead solely relying on git to readout the git hash. If not possible
then we resort to adding the directory as a "safe directory" and try
again.

Related progress issue: https://progress.opensuse.org/issues/113030