Add test for git version reading as non-owner
We already fixed the case when we try to read out a git version as a
different user than the owner of git repositories but we could not add
complete tests. This is my approach so far checking if we can use sudo
and use a user "nobody" which at least exists in a normal openSUSE
environment. Then the test injects a prefix "sudo -u nobody" into the
git version readout function.

Alternatives considered:
* https://metacpan.org/pod/Linux::Unshare but I failed to get it to work
  and it is not within our base distribution so far
* The command "unshare --user" which does not seem to reproduce the
  problem we wanted to fix

Related progress issue: https://progress.opensuse.org/issues/113030