tools: Exclude unreliable t/data/tests/tests from coverage
The coverage of our "openQA test modules" can not be directly collected
same as other perl code due to the nature of dynamically evaluating the
perl code modules within actual test runs. Also there is test code like
post_fail_hooks and such that is not necessarily meant to run at all
during normal tests so it's a better approach to explicitly exclude the
complete path for those modules.