Fix nr. of keys sent in send_key_until_needlematch
the function was sending the key n+1 times before failing. In test 03-testapi.t , key 'esc' was sent two times and then checked for three times.

This is an issue  for some scenarios, for example if an even number of keypresses is required, as is the case of maximizing and unmaximizing a window, by sending an even number of alt-f10.