Tune default similarity level for `wait_screen_change` when typing
At least when enabling `wait_screen_change` in os-autoinst's typing test
(`t/data/tests/tests/typing.pm` in this repo) the default similarity level
of 50 is not a good choice because especially when just typing a white
space (which only moves the blinking cursor) the test will struggle to
continue. So for typing letter-by-letter (which can be white-spaces) it
likely makes sense to increase the similarity level. With 60 the test runs
fine so I've been choosing that value.

With this change it is also possible to override the default similarity
level via `wait_screen_change_similarity_level`.