Fix resolution with virtio backend on new QEMU versions
The QEMU version 7.0.0 (currently in Tumbleweed) apparently changed the
default EDID information when using the virtio backend so now the screen
resolution is wrong once the graphical installer is shown. (The video
parameter usually specified within GRUB only affects the console but the
installer GUI makes its own EDID lookup.)

This change fixes the problem at least when `QEMUVGA` is set manually to
`virtio`.

Note that `cirrus` doesn't seem to be affected.