Consolidate qemu video device setting, deprecate QEMUVGA
As discussed in #2170, this code has progressively become a bit
of a mess as special cases were added for different arches and
things changed in qemu. This rewrites it to at least consolidate
the mess in one place, and deprecate some things which will
allow us to simplify it in future.

Always setting `-device` rather than sometimes `-vga` and
sometimes `-device` allows us to be much more consistent. We
'translate' `QEMUVGA` values into `-device` values after warning
that it is deprecated. Eventually we can hopefully get rid of
the handling of the deprecated vars, and the code will then be
much clearer and simpler.

References for this:
https://www.kraxel.org/blog/2019/09/display-devices-in-qemu/
https://www.kraxel.org/blog/2019/03/edid-support-for-qemu/

Signed-off-by: Adam Williamson <awilliam@redhat.com>