testapi: Try to return sane default in non-qemu host_ip calls
In case of non-qemu tests rely on host_ip, e.g. for log uploading. For
this so far a valid WORKER_HOSTNAME would need to be provided. We can
provide a sane default what is supplied by "hostname -f" which should
return a valid FQDN but can still be overridden with the variable
WORKER_HOSTNAME as necessary.

Related progress issue: https://progress.opensuse.org/issues/120261