Resume directly after failure in case `assert_script_run` fails
With this change, resuming the test execution ignoring failures will resume
the test execution right after the failing `assert_script_run` command
(instead on the next module). So far this is only a tweak for
`assert_script_run` but it could be extended for other test API functions.

I've been starting with `assert_script_run` because this function was the
starting point for the whole "pause on failure" feature.

Related ticket: https://progress.opensuse.org/issues/119380