Fix retry logic for git clone
The former code was not working properly as the return codes from the
clone_git method was different than expected. Also the test did not
completely agreed with the expectation.

Fix the return codes to be 1 on success everywhere, wrap the clone_git
into eval to catch the failures and retry in a loop.

References:
- https://progress.opensuse.org/issues/123873
- https://progress.opensuse.org/issues/123556
- https://progress.opensuse.org/issues/123888