Allow debugging JSON RPC per job
Occasionally it appears that the wait_serial command is sent by the
testapi frontend, but never received by the backend. This has only
been observed on a single job after a single LTP test.

It is not easily reproducible and so we need to start debugging in
production.

We need to confirm or rule out that the JSON RPC is being
lost. Meanwhile we do not want to enable debugging for all jobs that a
worker runs.

With this change we can enable debugging JSON RPC on a single job.

https://progress.opensuse.org/issues/121774