Fall back to manual fetch when checking out branch/ref of Git repo
We have already code in-place that is acting as fallback if `git clone`
with `--branch…` does not work. However, it seems rather complicated and
doesn't handle the case when the branch is actually just a ref like
`refs/pull/<pr_number>/merge` which is used by GitHub actions¹. This change
adds a new fallback (before the current one) that simply fetches the
missing ref and checks it out. If that fails the previous fallback is still
attempted because I'm not sure whether it might still be better in certain
cases.

Related issue: https://progress.opensuse.org/issues/124502

---

¹ See documentation of `GITHUB_REF` on
https://docs.github.com/en/actions/learn-github-actions/variables. Note
that using `GITHUB_REF_NAME` instead does *not* help as it is just a ref
like `<pr_number>/merge` which our current fallback code cannot resolve as
well.