Implement OCR based on reference string
The previous implementation of the OCR based matching did not compare
recognized text of the image area with a reference string.
This commit replaces the Tesseract implementation with an
implementation based on the program GOCR. With this commit, openQA
does not require OCR needles to have images provided along with the
JSON files. Instead, the reference string can be delivered in the
JSON. If the strings do not match, the Levenshtein distance is used
to calculate the match percentage.
Implicitly, this commit fixes a bug that prevented running openQA if
JSON files without images were provided.