Rename variable "scriptdir" for clarification
The top directory of os-autoinst was referenced as "scriptdir" even
though most references do not read actual "scripts" from there but other
data or modules hence renaming to "topdir" to clarify. This allows to
move scripts to another place to where we can then reference with a
different variable "scriptdir".