ci: Add slash to directories in codecov.yml
Issue: https://progress.opensuse.org/issues/125237

Apparently if one only specifies `t` then it counts all files
starting with t including testapi.pm and tools/update-deps etc.

It would be really good if there was a page on the codecov report where
it says which files it considered for the several groups.