ci: Improve marking uncoverable statements
It's only possible to find out which statements are covered or not by
having a close look at the HTML report of each file, especially at the
number in the "stmt" column.
This will also tell you how many statements a line actually has,
as it will add the number of calls in an extra row per statement.

Also if a statement is covered and marked as uncovered, it will decrease
the overall percentage (look for lines which are marked as uncoverable
but have a call number of greater than 0.
So I removed some of the uncoverable markers and added some elsewhere,
where one or more statements on a line aren't covered.
Now locally I get 100% for all t/*.t files.

Issue: https://progress.opensuse.org/issues/125237