Drop setting of TESSDATA_PREFIX in invoke-tests
It's not appropriate to do this here. This is not the right path
on all distros. Forcing this path here makes it impossible to run
the tests successfully on e.g. Fedora, where the tesseract data
is in /usr/share/tesseract/tessdata (and there is no need to
explicitly specify this, as tesseract is built to find it there
without any such specification).

Signed-off-by: Adam Williamson <awilliam@redhat.com>