dist: Fix unresolved dependency on chattr in OBS
https://build.opensuse.org/package/show/devel:openQA/isotovideo-qemu-x86-jq
shows that os-autoinst-qemu-x86 is unresolvable due to "nothing provides
/usr/bin/chattr". As explained by DimStar zypper would have no problem
to resolve but for a build project in OBS the project config would need
to define a "FileProvides". As the build project config is not
automatically synced for different build products we likely have a more
scalable solution with specifying the package that supplies chattr
directly.