Enable VNC endian conversion when worker is big-endian
The former approach only does the endian conversion when
no_endian_conversion attribute not defined or values of
$server_is_big_endian and $client_is_big_endian is not equal.

The $client_is_big_endian is not represents of the endiannes of the
WebUI and in fact they both returns same value depending of the
endianess of the Worker. This approach falsely don't do the endian
conversion when worker is big-endian, for instance on s390x.

The new approach does the endian conversion when $client_is_big_endian
and $client_is_big_endian returns true as a value.

Signed-off-by: Elkhan Mammadli <elkhan.mammadli@protonmail.com>