Provide network boot options for qemu
With this PR I first enable the option to pass network boot with
BOOTFROM. Qemu, however, provides `bootindex` property on the individual block
or net devices. This property is more flexible from `-boot`, where some aarch
do not support. For more read
https://www.qemu.org/docs/master/system/bootindex.html

I do not rethink and I did not refactor the entire logic, to prioritize
`bootindex` for _devices_ or _drives_. I merely enable this option in a
particular case (PXE) where the test requests to boot from network. As for
now, when PXEBOOT is given, and only on x86_64, the test constructs the qemu
command with `-boot n`. This option is not available on other archs like
aarch4 where the net boot is not doable.

Signed-off-by: ybonatakis <ybonatakis@suse.com>