Switch OBS SCM/CI Workflow to branch_package
The OBS changed how it decides if a service should run for a package. Services in modes that would not run server side are filtered out.

This means packages created with `link_package` steps stopped running services a while ago. 

Using the `branch_package` step with `add_repositories: disabled` is a replacement for having a package with services but only selected repositories in the test build.