Prevent external tool "update-deps" leaking into coverage analyis
tools/update-deps is a symlink to a script in external/ which is already
excluded from coverage analysis as it comes from an external repository.
tools/update-deps should hence also not be included in our coverage
analysis – especially as its statement coverage is not 100% – and as
tools/update-deps is currently the only file within tools/ that ends up
in coverage data anyway as visible in
https://app.codecov.io/gh/os-autoinst/os-autoinst/tree/master/tools
we should consistently exclude the complete tools directory from
analysis.