Avoid CMake deprecation warning by bumping version
Avoid warnings like

```
CMake Deprecation Warning at ppmclibs/CMakeLists.txt:2 (cmake_minimum_required):
  Compatibility with CMake < 3.5 will be removed from a future version of
  CMake.

  Update the VERSION argument <min> value or use a ...<max> suffix to tell
  CMake that the project does not need compatibility with older versions.
```

after the recent CMake update on Tumbleweed by bumping the
version to 3.17.0 which is the CMake version provided by
Leap 15.3 (so this change should be ok with regard to
compatibility).

It looks like we don't rely on any legacy behavior of that old
CMake version so not other changes are required.