os-autoinst-openvswitch: Fix running into timeout for slow network init
In case of many openQA worker instances where each instance has one to
multiple TAP devices assigned the network initialization can take
multiple minutes which has been observed in particular when using the
"wicked" network manager. We should account for that and just wait
longer. This is what we did for the openqa.opensuse.org and other openQA
instances already for years with a systemd override setting the
environment variable.

This commit bumps the timeout already within os-autoinst-openvswitch
from 1 to 5 minutes to account for that.

Related progress issue: https://progress.opensuse.org/issues/132134