Remove wrong condition for `unxz` call on svirt host
The check for the existence of the unextracted asset is done one
the wrong host. The check was supposedly meant to be an
optimization but it never worked. So it is better to remove it
completely.

This condition would have also prevented the extraction in case
an unextracted version was present on the worker host.