Simplify log message when copying asset to svirt host
Just log the filename (e.g. `foo.qcow`) and not the full path (e.g.
`/var/lib/openqa/share/factory/hdd/fixed/foo.qcow2`) because the full
path might actually not be correct and therefore misleading. It might
as well be an equally named asset in the non-fixed directory that was
actually synced by the openQA worker (which in fact has a preference for
syncing non-fixed assets if the same asset exists in both directories).

In this context it isn't clear to which host this path relates to
anyways.