Consider code as generally uncoverable when testing relies on `sudo`
It is generally not good to rely on the existence of a specific user and
that we can execute a command as that user via `sudo` in our unit tests.
The tests are therefore already conditional. This change takes care that we
don't get different coverage results depending on that condition.