video_stream: Add ustreamer support
ustreamer is a very efficient application for reading /dev/video0 and
sending the video stream in various ways. It's part of PiKVM project and
is optimized to work on RPi with HDMI input specifically. One of the
interfaces is providing raw frames over a shared memory sink (via
POSIX shared memory) and that looks like the most useful interface for
openQA worker.

Add support for ustreamer, including:
- starting the ustreamer process with appropriate args
- accessing shared memory, including synchronization with ustreamer via
  flock()
- decoding both UYVY frames (received from v4l2) and JPEG (static frame
  "no signal")

Since ustreamer handles DV timings itself, skip this part on the
os-autoinst side.