VNC: add support for Tight encoding
This extends supported VNC servers, especially it can be used with
the one from PiKVM (useful when openqa-worker is running on a different
host than PiKVM, otherwise ustreamer will give better results).

This implementation supports JPEG and Fill compression, which is enough
to serve PiKVM (only JPEG used) and Xvnc (both Fill and JPEG used).
BasicCompression is not implemented here.
To make lossless encodings preferred, place them manually at the
beginning of the list, instead of sorting by value. But also, disable it
by default, since by the spec it should be preferred over RAW (but not
over ZRLE). To enable it with generalhw backend, set
GENERAL_HW_VNC_JPEG=1.

Similar implementation (if not exactly the same) could be used for
implementing Tight PNG encoding (-260).
JPEG encoding (21) is a bit different, as there is no explicit size
(parser needs to read data until EOF marker), and it may re-use Huffman
tables.
Neither of the last two is needed for PiKVM, so this commit do not
implement them.