Add UEFI boot support for vmware tests (poo#152917)
Due to UEFI becoming the mainstream system boot option for user OS
installation, it should also support UEFI boot in VMware tests.
From VMware ESXi 7.0 guest installation, the boot option switches
from BIOS to UEFI as default, and it will help to workaround bsc#1216655.
For VMware guests, the firmware attribute is set to efi when the guest
uses UEFI, and it is not set when using BIOS.

The libvirt XML format described as below.

- Legacy boot (BIOS)
  <os>
    <type>hvm</type>
  </os>

- UEFI boot
  <os firmware='efi'>
    <type>hvm</type>
  </os>