script_run: base serial buffer size on command length
If you use script_run at a serial console to run a command more
than 4096 characters long, the wait_serial that waits until the
whole command has made it across the serial connection will
fail, because it uses the default buffer size (4096), so it
will never see the full command. This means there'll be a 90
second timeout before the command is actually run, and a red
frame in the results.

This solves the problem by always sizing the buffer to the length
of the command, plus 128 extra bytes (for the marker string, and
some more for safety).

Issue: https://progress.opensuse.org/issues/153304

Signed-off-by: Adam Williamson <awilliam@redhat.com>