video_stream: specify preferred video format explicitly
ustreamer's raw sink outputs whatever v4l2 device returns. Without
explicit option, it will use the format that was configured there
before. Normally this is UYVY for the tc358743 driver (that handles HDMI
input on Raspberry Pi), but other application could change it in the
meantime. Avoid breaking the test in such case and request the UYVY
format explicitly.