Unify size argument for disk creation
There was an inconsistency between _copy_image_to_vm_host and add_disk
where one assumed the size to be an int and the other method assumed to
be in the format qemu-img expects it (e.g. 20G).

This commit unifies the argument by making it everywhere use the
qemu-img format, as it is used widely in the os-autoinst-distri-opensuse
code.