ustreamer: restrict to aarch64 and x86_64
The use of `unpack` in the ustreamer code assumes a 64-bit,
little-endian environment. It fails on s390x (because it's big-
endian) and i686 (because it's 32-bit). In practice, ustreamer is
really only used on aarch64. Let's bail unless we're running on
aarch64 or x86_64, and similarly disable the tests outside of
those arches. We keep x86_64 for the convenience of testing in
CI environments and on developers' local machines.

Related progress issue: https://progress.opensuse.org/issues/161969

Signed-off-by: Adam Williamson <awilliam@redhat.com>