Reliably locate top directory of isotovideo
Before we matched (/some/path/to/)isotovideo to get the top directory.
In the future we might also need to support /some/path/to/scripts/isotovideo
being symlinked.

When we move isotoideo to scripts/ and symlink it in the top directory,
both invocations need to work.

For example, openQA's fullstack test is using:

    perl script/worker --isotovideo=../os-autoinst/isotovideo --verbose

And OpenQA::Worker::Engines::isotovideo is using
abs_path(../os-autoinst/isotovideo) which will return the path to the linked
script, not the symlink itself.

So we should get the real path of the script, which $FindBin::RealBin is
realiably doing, and then add '..' to go to the top.