Remove obsolete support for "script_run" with "die_on_timeout=0"
The correct and expected way to call script_run is

```
script_run($cmd [, timeout => $timeout] [, output => ''] [, quiet => $quiet])
```

which will die on timeout. The "die on timeout" is the same behaviour as for
other commands like assert_screen and such. As the old behaviour had the
deprecation message in place for more than one year now we should remove
support for the obsolete way.

The idea of `script_run` is to await the execution of a command and return the
result. If one does not care about the result and does not want to wait for the
execution to be done then one can still just use `enter_cmd`. This would only
be a problem if any subsequent command can not handle that the previous one did
not finish. As alternative consider

```
script_run 'timeout $lower_timeout $cmd', timeout => $higher_timeout
```

Another but not suggested way would be

```
eval { script_run($cmd) };
if ($@) { $handle_timeout }
```

Related progress issue: https://progress.opensuse.org/issues/162035