video_stream: make FPS configurable
This is all about balance between compute power on the worker host and
desired latency and video smoothness. Ustreamer is a bit lighter than
ffmpeg, so it can capture more frames (on Raspberry Pi), but then it
puts more stress on the system as more frames are searched for needles
and more frames goes to video encoder. On the other hand, lower FPS may
cause some details to be missed (like quickly dismissed notification).
The optimal value may even differ from test to test, for example based
on the test length and how much backlog for video encoder is acceptable
(amount of available RAM is important in this case too).
Instead of trying to find perfect value that fits all, make it
configurable. Keep the default unchanged.