VMWare: add combustion and ignition support
In VMWare's `vmx` we can pass provisioning information for the current
VM. The configuration data in case of cloud-init or ignition as well as
configuration script, in case of combustion has to encoded in base64 at
least. Usually users would compress and encode the content of these
files. It would be user-friendly if we can pass only the file name that
should be located in osado's `data` folder