Avoid dependency on `File::Touch` for easier SLE submission
We are now using `Mojo::File` almost everywhere anyway so let's avoid one
last usage of the `File::Touch` module. This means we can avoid submitting
`File::Touch` to SLE which would be problematic because the current version
of that module depends on a version of a Perl core module that is too old
on the SLE-provided Perl (`perl(Time::HiRes) >= 1.976`).

Related ticket: https://progress.opensuse.org/issues/128318